Hackbot/requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
Deprecated==1.2.11
discord.py==1.6.0
dnspython==2.1.0
idna==2.10
multidict==5.1.0
PyGithub==1.54.1
PyJWT==1.7.1
pymongo==3.11.2
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
yarl==1.6.3
